+++ stroom-app/src/main/java/stroom/startup/App.java
@@ -34,7 +34,6 @@ import org.eclipse.jetty.servlet.ServletHolder;
 import org.eclipse.jetty.servlets.CrossOriginFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.BeanCreationException;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import stroom.cluster.server.ClusterCallServiceRPC;
@@ -82,20 +81,7 @@ import stroom.servicediscovery.ResourcePaths;
 import stroom.servicediscovery.ServiceDiscovererImpl;
 import stroom.servicediscovery.ServiceDiscoveryManager;
 import stroom.servicediscovery.ServiceDiscoveryRegistrar;
-import stroom.servlet.CacheControlFilter;
-import stroom.servlet.DashboardServlet;
-import stroom.servlet.DebugServlet;
-import stroom.servlet.DynamicCSSServlet;
-import stroom.servlet.EchoServlet;
-import stroom.servlet.ExportConfigServlet;
-import stroom.servlet.HttpServletRequestFilter;
-import stroom.servlet.ImportFileServlet;
-import stroom.servlet.RejectPostFilter;
-import stroom.servlet.SessionListListener;
-import stroom.servlet.SessionListServlet;
-import stroom.servlet.SessionResourceStoreImpl;
-import stroom.servlet.StatusServlet;
-import stroom.servlet.StroomServlet;
+import stroom.servlet.*;
 import stroom.spring.MetaDataStatisticConfiguration;
 import stroom.spring.PersistenceConfiguration;
 import stroom.spring.ScopeConfiguration;
@@ -278,7 +264,6 @@ public class App extends Application<Config> {
         SpringUtil.addServlet(servletContextHandler, applicationContext, ImportFileServlet.class, ResourcePaths.ROOT_PATH + "/importfile.rpc");
         SpringUtil.addServlet(servletContextHandler, applicationContext, ScriptServlet.class, ResourcePaths.ROOT_PATH + "/script");
         SpringUtil.addServlet(servletContextHandler, applicationContext, ClusterCallServiceRPC.class, ResourcePaths.ROOT_PATH + "/clustercall.rpc");
-        SpringUtil.addServlet(servletContextHandler, applicationContext, ExportConfigServlet.class, ResourcePaths.ROOT_PATH + "/export");
         SpringUtil.addServlet(servletContextHandler, applicationContext, StatusServlet.class, ResourcePaths.ROOT_PATH + "/status");
         SpringUtil.addServlet(servletContextHandler, applicationContext, EchoServlet.class, ResourcePaths.ROOT_PATH + "/echo");
         SpringUtil.addServlet(servletContextHandler, applicationContext, DebugServlet.class, ResourcePaths.ROOT_PATH + "/debug");
@@ -293,6 +278,7 @@ public class App extends Application<Config> {
         SpringUtil.addServletListener(environment.servlets(), applicationContext, SessionListListener.class);
 
         // Add resources.
+        SpringUtil.addResource(environment.jersey(), applicationContext, ExportConfigResource.class);
         SpringUtil.addResource(environment.jersey(), applicationContext, DictionaryResource.class);
         SpringUtil.addResource(environment.jersey(), applicationContext, RuleSetResource.class);
         SpringUtil.addResource(environment.jersey(), applicationContext, StroomIndexQueryResource.class);