+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Uplift send_to_stroom.sh in the distribution to v2.0
 
+* Issue **#973** : Export servlet changed to a Resource API, added permission check, improved error responses.
+
 ## [v6.0-beta.17] - 2018-12-04
 
 * Issue **#969** : The code now suppresses errors for index shards being locked for writing as it is expected. We now lock shards using maps rather than the file system as it is more reliable between restarts.